ragMainProject/requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.0
django-cors-middleware==1.5.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
Pillow==8.4.0
pip==21.3.1
PyJWT==2.3.0
pytz==2021.3
setuptools==58.5.3
six==1.16.0
sqlparse==0.4.2
